+++ packages/eyes-sdk-core/lib/TestResults.js
@@ -342,7 +342,7 @@ class StepInfo {
  * @typedef TestAccessibilityStatus
  * @type {object}
  * @property {AccessibilityLevel} level - accessibility level.
- * @property {AccessibilityVersion} version - accessibility version.
+ * @property {AccessibilityGuidelinesVersion} version - accessibility guidelines version.
  * @property {AccessibilityStatus} status - test accessibility status.
  */