+++ packages/eyes-selenium/test/coverage/custom/TestVGServerConfigs.spec.js
@@ -4,7 +4,7 @@ const {
   Configuration,
   BrowserType,
   AccessibilityLevel,
-  AccessibilityVersion,
+  AccessibilityGuidelinesVersion,
   DeviceName,
   MatchLevel,
 } = require('../../../index')
@@ -53,7 +53,7 @@ describe('TestVGServerConfigs', () => {
     conf
       .setAccessibilityValidation({
         level: AccessibilityLevel.AA,
-        version: AccessibilityVersion.WCAG_2_0,
+        version: AccessibilityGuidelinesVersion.WCAG_2_0,
       })
       .setIgnoreDisplacements(false)
     eyes.setConfiguration(conf)
@@ -63,7 +63,7 @@ describe('TestVGServerConfigs', () => {
     conf
       .setAccessibilityValidation({
         level: AccessibilityLevel.AAA,
-        version: AccessibilityVersion.WCAG_2_1,
+        version: AccessibilityGuidelinesVersion.WCAG_2_1,
       })
       .setIgnoreDisplacements(true)
     eyes.setConfiguration(conf)
@@ -73,7 +73,7 @@ describe('TestVGServerConfigs', () => {
     conf
       .setAccessibilityValidation({
         level: AccessibilityLevel.AA,
-        version: AccessibilityVersion.WCAG_2_0,
+        version: AccessibilityGuidelinesVersion.WCAG_2_0,
       })
       .setMatchLevel(MatchLevel.Layout)
     eyes.setConfiguration(conf)
@@ -89,7 +89,7 @@ describe('TestVGServerConfigs', () => {
       await assertDefaultMatchSettings(result, {
         accessibilitySettings: {
           level: AccessibilityLevel.AA,
-          version: AccessibilityVersion.WCAG_2_0,
+          version: AccessibilityGuidelinesVersion.WCAG_2_0,
         },
         ignoreDisplacements: false,
         matchLevel: MatchLevel.Strict,
@@ -97,7 +97,7 @@ describe('TestVGServerConfigs', () => {
       await assertImageMatchSettings(result, {
         accessibilitySettings: {
           level: AccessibilityLevel.AAA,
-          version: AccessibilityVersion.WCAG_2_1,
+          version: AccessibilityGuidelinesVersion.WCAG_2_1,
         },
         ignoreDisplacements: true,
         matchLevel: MatchLevel.Strict,
@@ -107,7 +107,7 @@ describe('TestVGServerConfigs', () => {
         {
           accessibilityLevel: {
             level: AccessibilityLevel.AA,
-            version: AccessibilityVersion.WCAG_2_0,
+            version: AccessibilityGuidelinesVersion.WCAG_2_0,
           },
           ignoreDisplacements: true,
           matchLevel: MatchLevel.Layout2,