+++ packages/eyes-sdk-core/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 
 ## Unreleased
+
+## 9.2.0 - 2020/4/26
+
 - updated to @applitools/eyes-common@3.22.0
 
 ## 9.1.2 - 2020/4/23